Route added + changes to the airplane when it comes to its flight position
A route can be added with multiple routeLines. The route.setAirplaneNextPosition can be called to move the airplane to its next position on that route. This was handled in the airplane class itself, but since it can be on one of the multiple routeLines it is easier to handle this from the route class itself.