RouteLine take-off and landing improved
The RouteLine will remove the first and last points of the line at flight-height untill the distance from airport to the first point is at least the sqrt of 2*(fleightHeight^2). This causes an angle of 45 degrees at max.
If the distance between airports is not long enough to reach its travel height it will only create 3 points and make a curve out of those.
Apart from that there have been some improvements to optimise the code