low-poly model of plane + coloured textures instead of existing company textures
Vertices reduced from 3787 to 644.  Edges reduced from 6469 to 1267. Faces reduced from 2860 to 629. Tris reduced from 5626 to 1204. UVs reduced from 6546 to 908.
This process caused the UV map to be impossible to work with. Therefore the high-poly and low-poly models both need their own textures.
The textures' hue and saturation are changed to get the colours.
(Shaders could be used as well. This would eliminate the need for this many textures, and it gives the possibility to get an infinite number of colours. This would require 2 layers of texutres since the coloured layer needs to be changed without affecting the white base-colour of the plane.)