small fix
Due to previous changes the plane wouldn't scale up and down. This has been fixed.
A bug has been found: when removing points from the curve, the distance between points[0] and points[1] increases. this causes a big difference in distance between points[0] to points[1], and points[1] to points[2]. When moving the plane it will takes exactly the same amount of time to move from points[0] to points[1] as it is to move from points[1] to points[2]. Therefore the speed (distance/time) will differ quite a lot.