Airport looks + selection + hover update
An airport is now shown as a red circle instead of pink box.
You can hover over an airport to make it bigger.
You can click on an airport to select it, which makes it blue.
If you want to create a new Route, the selected airport turns green, and stays green for as long as it is in the selection.
If you want to create a new Route, and the airport you click on is already in the list of selected airports it will now remove it from the list and de-select it.