bugfix airport selection
Fixed a bug when the selected airport (blue) were to be selected again after 'new route' selection had been turned on. This used to put the airport in the list of airports that would be added to the route, but the airport object didn't become green, but reverted back to the default red colour.

Fixed a bug where the airports kept their green colour after a new route had been created.