Small optimalisation tweak
The newRoute array had an array of object.info._id. The newRouteObjs array had an array of object. Since object already has object.info._id inside of it there is no need to have a seperate array with object.info._id. Therefore said array has been removed and any reference to the id's it contained have been changed.
The temporary visual new route used the object.info._id to create an array of objects by going through the list of all airports to check if the ids match. Since we now already have an array with the objects there is no need to do this, and therefore said code has been removed