RouteLine visual bug fixes
Fixed a bug where the short routes could get a line which would be higher than the flightHeight.
Fixed a bug where the short routes could get a line which would have a steeper increase than the longer lines.
Fixed a bug where the plane would fly along a different (the correct) path.