Airplane scaling improved
If the highest point of a routeLine was 6, and the maximum height a line can be was 10, the airplane's scale on the highest point would be 6/10=0.6.
Now it calculates the highest point of the routeLine (on creation) and uses that value as its max value. This way it calculates its scale on the highest point like 6/6=1.