airports and routes can be deselected
A selected airport of route is now deselected when the user clicks on something else within the scene.
glitch found: A selected route won't be rendered when it is directly in front of the background. The parts of the route which are behind or in front of any other object are visible.