routeList html now updates + bug fix after adding route + axis helper removed
The axis helper for debugging purposes has been removed since it is now irrelevant.
There was a bug where the 'new route' button was set back to the old button type, which caused it to stop working. This is now fixed.
The list of current routes (bottom right of the screen) didn't update once a new route was added. This is now fixed.